move assertion helpers to test suite
