+++ pkg/cmd/config/set_garden_test.go
@@ -12,7 +12,6 @@ import (
 	. "github.com/onsi/gomega"
 	"github.com/onsi/gomega/types"
 	"github.com/spf13/cobra"
-	"github.com/spf13/pflag"
 
 	cmdconfig "github.com/gardener/gardenctl-v2/pkg/cmd/config"
 	"github.com/gardener/gardenctl-v2/pkg/config"
@@ -137,36 +136,3 @@ var _ = Describe("Config Subcommand SetGarden", func() {
 		})
 	})
 })
-
-func assertAllFlagNames(flags *pflag.FlagSet, expNames ...string) {
-	actNames := []string{}
-
-	flags.VisitAll(func(flag *pflag.Flag) {
-		actNames = append(actNames, flag.Name)
-	})
-
-	ExpectWithOffset(1, actNames).To(Equal(expNames))
-}
-
-func assertGardenNames(cfg *config.Config, names ...string) {
-	ExpectWithOffset(1, cfg.GardenNames()).To(Equal(names))
-}
-
-func assertGarden(cfg *config.Config, garden *config.Garden) {
-	g, err := cfg.Garden(garden.Name)
-	ExpectWithOffset(1, err).ToNot(HaveOccurred())
-	ExpectWithOffset(1, g).To(BeEquivalentTo(garden))
-}
-
-func assertConfigHasBeenSaved(cfg *config.Config) {
-	c, err := config.LoadFromFile(cfg.Filename)
-	ExpectWithOffset(1, err).NotTo(HaveOccurred())
-
-	for i, g := range cfg.Gardens {
-		if len(g.Patterns) == 0 {
-			cfg.Gardens[i].Patterns = nil
-		}
-	}
-
-	ExpectWithOffset(1, c).To(BeEquivalentTo(cfg))
-}
